Port core to JS and Native
