+++ .github/workflows/ci.yml
@@ -49,7 +49,7 @@ jobs:
       matrix:
         java: ['adopt@1.8', 'adopt@1.11']
         scala: ['2.11.12', '2.12.13', '2.13.5', '3.0.0-RC3']
-        platform: ['JVM']
+        platform: ['JS', 'JVM', 'Native']
     steps:
       - name: Checkout current branch
         uses: actions/checkout@v2.3.4
