+++ core/js/src/main/scala/zio/config/PropertyTypePlatformSpecific.scala
@@ -1,0 +1,3 @@
+package zio.config
+
+trait PropertyTypePlatformSpecific {}
