+++ core/native/src/main/scala/zio/config/ConfigDescriptorPlatformSpecific.scala
@@ -1,0 +1,3 @@
+package zio.config
+
+trait ConfigDescriptorPlatformSpecific {}
