+++ core/shared/src/main/scala/zio/config/PropertyType.scala
@@ -2,8 +2,6 @@ package zio.config
 
 import zio.config.PropertyType.PropertyReadError
 
-import java.io.File
-import java.net.{URI, URL}
 import java.time.{Instant, LocalDate, LocalDateTime, LocalTime}
 import java.util.UUID
 import scala.concurrent.duration.Duration
@@ -14,7 +12,7 @@ trait PropertyType[V, A] {
   def write(a: A): V
 }
 
-object PropertyType {
+object PropertyType extends PropertyTypePlatformSpecific {
   final case class PropertyReadError[V](value: V, typeInfo: String)
 
   case object StringType extends PropertyType[String, String] {
@@ -80,12 +78,6 @@ object PropertyType {
     def write(value: BigDecimal): String                                   = value.toString
   }
 
-  case object UriType extends PropertyType[String, URI] {
-    def read(value: String): Either[PropertyReadError[String], URI] =
-      attempt(new URI(value), _ => PropertyReadError(value, "uri"))
-    def write(value: URI): String                                   = value.toString
-  }
-
   case object DurationType extends PropertyType[String, Duration] {
     def read(value: String): Either[PropertyReadError[String], Duration] =
       attempt(Duration.apply(value), _ => PropertyReadError(value, "duration"))
@@ -128,28 +120,7 @@ object PropertyType {
     def write(value: Instant): String                                   = value.toString
   }
 
-  case object FileType extends PropertyType[String, File] {
-    def read(value: String): Either[PropertyReadError[String], File] =
-      attempt(new File(value), _ => PropertyReadError(value, "file"))
-
-    def write(value: File): String = value.toString
-  }
-
-  case object UrlType extends PropertyType[String, URL] {
-    def read(value: String): Either[PropertyReadError[String], URL] =
-      attempt(new URL(value), _ => PropertyReadError(value, "url"))
-
-    def write(value: URL): String = value.toString
-  }
-
-  case object JavaFilePathType extends PropertyType[String, java.nio.file.Path] {
-    def read(value: String): Either[PropertyReadError[String], java.nio.file.Path] =
-      attempt(java.nio.file.Paths.get(value), _ => PropertyReadError(value, " java.nio.file.Path"))
-
-    def write(value: java.nio.file.Path): String = value.toString
-  }
-
-  private def attempt[A, E](a: => A, f: Throwable => E): Either[E, A] =
+  protected def attempt[A, E](a: => A, f: Throwable => E): Either[E, A] =
     Try(a) match {
       case Success(value)     => Right(value)
       case Failure(throwable) => Left(f(throwable))
